adapters/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dataspace-control-plane-adapters"
version = "0.1.0"
description = "Anti-corruption and integration layer — implements core/ ports for every external protocol, product, and runtime"
requires-python = ">=3.12"
dependencies = [
    "dataspace-control-plane-core",
    # HTTP
    "httpx>=0.27",
    # Async
    "anyio>=4.3",
    # Configuration / validation
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    # Database
    "asyncpg>=0.29",
    "sqlalchemy[asyncio]>=2.0",
    # Kafka
    "confluent-kafka>=2.4",
    # Keycloak / OIDC
    "python-jose[cryptography]>=3.3",
    # Temporal
    "temporalio>=1.7",
    # Vault
    "hvac>=2.3",
    # OpenTelemetry
    "opentelemetry-api>=1.24",
    "opentelemetry-sdk>=1.24",
    "opentelemetry-exporter-otlp>=1.24",
    # Retry / resilience
    "tenacity>=8.3",
    # Crypto (signing, DID, VC)
    "cryptography>=42.0",
    # Object storage
    "aiobotocore>=2.13",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-httpx>=0.30",
    "respx>=0.21",
]

[tool.hatch.build.targets.wheel]
packages = ["src/dataspace_control_plane_adapters"]
